slab: Dissolve package into object and worker
Also heavily refactor worker
